% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\name{internal}
\alias{internal}
\title{Documentation}
\description{
This page lists and describes some internal functions
of the R package \code{\link{semisup}}.
These functions should not be used for analysing data.

\code{\link{fit.wrap}} multiple restarts
\cr \code{\link{fit.norm}} Gaussian mixture model
\cr \code{\link{fit.nbinom}} negative binomial mixture model
\cr \code{\link{fit.zinb}} zero-inflated negative binomial mixture model
\cr \code{\link{estim.nbinom}} dispersion estimation
\cr \code{\link{estim.zinb}} dispersion and zero-inflation estimation
\cr \code{\link{resam.lrts}} resampling (bootstrap, permutation)
}
\seealso{
Use \code{\link{mixtura}} for model fitting,
and \code{\link{scrutor}} for hypothesis testing.
}
\keyword{internal}
