% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{semisup-package}
\alias{semisup-package}
\alias{semisup}
\title{Semi-supervised mixture model}
\description{
This R package implements the semi-supervised mixture model.
Use \code{\link{mixtura}} for model fitting,
and \code{\link{scrutor}} for hypothesis testing.
}
\section{Getting started}{

Please type the following commands\strong{:} \cr
\code{utils::vignette("semisup")} \cr
\code{?semisup::mixtura} \cr
\code{?semisup::scrutor}
}

\section{More information}{

A Rauschenberger, RX Menezes, MA van de Wiel,
NM van Schoor, and MA Jonker (2017).
"Detecting SNPs with interactive effects on a quantitative trait",
\emph{Manuscript in preparation}.

\email{a.rauschenberger@vumc.nl}
}

\keyword{documentation}
