% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{data}
\name{table}
\alias{table}
\title{Table}
\format{A list of numeric vectors.}
\usage{
data(table)
}
\value{
All entries are numeric.
}
\description{
This dataset includes tables for the approximate mixture test
(\strong{not yet available}).
}
\keyword{internal}
